auth/requirements.txt
fastapi==0.83.0
uvicorn==0.18.3
pymongo[srv]==4.2.0
pyjwt==1.7.1
